+++ docs/CHANGELOG.md
@@ -1,11 +1,29 @@
 # CHANGELOG
 
-# Version 1.0.1 2019-10-02
+## Version 1.1.0 2020-01-23
+
+- Update license year, happy 2020!
+- Include cancellation document for document *"CFDI de retención e información de pagos"*.
+- Add `DocumentType` enumerator with keys `cfdi` and `retention` to specify the correct namespace of the request.
+- Add `XmlCancellationHelper::signRetentionCancellation` and `XmlCancellationHelper::signRetentionCancellationUuids`
+  that create cancellation request for retentions.
+- Refactor `XmlCancelacionHelper` and delegate the creation of the `Cancellation` object to a specific
+  protected method.
+- Other capsules uses also `DocumentType` to set the main namespace but as hardcoded.
+- Development:
+    - Move from `phpstan/phpstan-shim` to `phpstan/phpstan`.
+    - Upgrade to `phpstan/phpstan: ^0.12`
+- Testing:
+    - Improve `CancellationTest` to check that `DocumentType` is used correctly.
+    - Create a testing class `XmlCancelacionHelperSpy` to spy on `XmlCancelacionHelper`.
+    - Refactor tests on `XmlCancelacionHelperTest` to test against the spy class.  
+
+## Version 1.0.1 2019-10-02
 
 - Fix documentation to point out version 1.0 instead of 0.5.
 - Fix documentation PHP examples.
 
-# Version 1.0.0 2019-09-28
+## Version 1.0.0 2019-09-28
 
 - This version is a major change, it is not compatible with previous versions
   Read [UPGRADE-1.0](https://github.com/phpcfdi/xml-cancelacion/blob/master/docs/UPGRADE-1.0.md)